pipe: wire display pixel format more generally
this also works with 8-bit export transparently. one caveat is that if you
connect the output of a write:* module to multiple inputs, you need to connect
the one which actually cares about the input format first.
unfortunately this also means if you route an image to 8-bit export *and* to 10-bit display,
it is currently an error (we could alternatively negotiate for 8-bit display, too, but
that sounds stupid).