Create draw-double-tint.pst
This is a preset that allows to change the color of specific areas of a photo in two different ways with the help of drawn masks/selections. Change e.g. saturation in colour tint1 and draw over an area, and change e.g. blue in colour tint2 and draw over an area.